feat:新增已通知数据

+++ app/src/main/java/net/ankio/vpay/utils/PayInfo.kt
@@ -1,3 +1,3 @@
 package net.ankio.vpay.utils
 
-data class PayInfo(val time: String, val type: Int, val money: Float)
+data class PayInfo(val time: String, val type: Int, val money: Double)
